+++ driver/src/main/java/oracle/nosql/driver/StatsControl.java
@@ -1,5 +1,5 @@
 /*-
- * Copyright (c) 2011, 2021 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2011, 2022 Oracle and/or its affiliates. All rights reserved.
  *
  * Licensed under the Universal Permissive License v 1.0 as shown at
  *  https://oss.oracle.com/licenses/upl/
@@ -82,6 +82,8 @@ import oracle.nosql.driver.values.MapValue;
  *
  * For a detailed statistics log entries structure and values see
  * {@link oracle.nosql.driver}</p>
+ *
+ * @since 5.2.30
  */
 public interface StatsControl {